pyproject.toml
[project]
name = "deinopis"
version = "0.1.0"
description = "Deinopis — KubeCon NA 2026 demo platform for 'Can Your Chatbot Run kubectl? Guardrails for LLMs on Kubernetes'"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["tests", "."]
timeout = 120
markers = [
    "phase1: Phase 1 — GKE Foundation",
    "phase2: Phase 2 — GitOps Bootstrap",
    "phase3: Phase 3 — The Eyes (observability)",
    "phase4: Phase 4 — The Net — Security (Kyverno + Falco)",
    "phase5: Phase 5 — The Net — AI Gateway",
    "phase6: Phase 6 — BurritBot application",
    "phase7: Phase 7 — Audience frontend",
    "phase8: Phase 8 — Hardening",
    "static: Tests that can run without a live cluster",
    "live: Tests that require a live cluster / GCP auth",
]

[dependency-groups]
dev = [
    "google-cloud-container>=2.50.0",
    "google-cloud-secret-manager>=2.20.0",
    "google-cloud-storage>=2.18.0",
    "kubernetes>=35.0.0",
    "pillow>=12.1.1",
    "pytest>=9.0.2",
    "pytest-timeout>=2.4.0",
    "pyyaml>=6.0.2",
    "qrcode>=8.2",
    "requests>=2.32.5",
]
